Get no of instructions in an optimised and better way. (1/2)
Earlier we just used a workaround by brute force a large no of instructions with no real way to check if the next instructions exist or not.
This implementation fixes it by seeing the output.

We didnt put an if condition for every instruction as it significantly increased the time required to execute. This is an optimised approach for the work.

Signed-off-by: AyushR1 <ayushrathore12501@gmail.com>