Squash: Get no of instructions in an optimised and better way.
Re add this the issue earlier was due to permisions.

Earlier we just used a workaround by brute force a large no of instructions with no real way to check if the next instructions exist or not.
This implementation fixes it by seeing the output.

We didnt put an if condition for every instruction as it significantly increased the time required to execute. This is an optimised approach for the work.

Squash of :-
 https://github.com/AyushR1/Dynamic-Instructions-Calculation/commit/fbed9ffb4656c85774426034c41dd31e4bba677b
https://github.com/AyushR1/Dynamic-Instructions-Calculation/commit/a8289fd662a751b057f38aaac92da283bdbb3441

Signed-off-by: AyushR1 <ayushrathore12501@gmail.com>